feature: first commit with db type and supporting bits.
The db type defines a type that wrappers the ORM (gorm).  Along with
this initial type, we had to include test support functions and
that necessitated including options plus errors as well.  At this
point some of the options are premature, but much of this pkg is
being brought over from Boundary. Be patient, in a commit or two all
these options will make sense.