update gorm dependencies (#23)
* feature: validate resource the correct "kind"

These changes are required to update the gorm dependency.

In the latest update, gorm appears to have become more sensitive when
a slice is passed into operations.  The ptr to slice must now
be a ptr to slice of ptrs.

* chore: update gorm dependencies

 Updating the gorm dependencies included updating its gorm.io/driver
dependencies

Gorm had a breaking change on the Field.ValueOf func which now
requires a context.  Thankfully, this breaking change did not leak
outside of the go-dbw package.

* fix: ensure operations honor/use passed in context